+++ mllib/src/main/scala/org/apache/spark/mllib/tree/impurity/Gini.scala
@@ -133,7 +133,7 @@ private[spark] class GiniCalculator(stats: Array[Double]) extends ImpurityCalcul
   /**
    * Number of data points accounted for in the sufficient statistics.
    */
-  def count: Long = stats.sum.toLong
+  def count: Double = stats.sum
 
   /**
    * Prediction which should be made based on the sufficient statistics.
